+++ vm/zone.c
@@ -43,8 +43,7 @@ int _vm_zoneCreate(vm_zone_t *zone, size_t blocksz, unsigned int blocks)
 		return -ENOMEM;
 	}
 
-	if ((zone->vaddr = vm_mmap(zone_common.kmap, zone_common.kmap->start, zone->pages, 0x1U << zone->pages->idx, PROT_READ | PROT_WRITE, zone_common.kernel, -1, MAP_NONE)) == NULL) {
-		// TBD_Julia teraz to 0x1U ma za duży rozmiar - idx  jest u8
+	if ((zone->vaddr = vm_mmap(zone_common.kmap, zone_common.kmap->start, zone->pages, (size_t)0x1U << zone->pages->idx, PROT_READ | PROT_WRITE, zone_common.kernel, -1, MAP_NONE)) == NULL) {
 		vm_pageFree(zone->pages);
 		return -ENOMEM;
 	}
